refactor
the Database is not specific to the project and therefore not needed in the project folder.
