+++ README.md
@@ -1,6 +1,2 @@
 # gradeCalculator
 Allows the user to enter his grades, save them and see what the average, best or worst end grade will be.
-
-start Database: "docker compose up"
-connect to Database: "mysql -P 8083 --protocol=tcp -u root -p"
-start Programm: ./gradlew bootRun
