+++ src/main/java/com/gradeCalculator/ServerApplication.java
@@ -1,4 +1,4 @@
-package com.gradeCalculator.server;
+package com.gradeCalculator;
 
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.EnableAutoConfiguration;
