+++ src/main/java/com/gradeCalculator/repositories/ModuleRepository.java
@@ -1,7 +1,7 @@
-package com.gradeCalculator.server.repositories;
+package com.gradeCalculator.repositories;
 import org.springframework.data.repository.CrudRepository;
 
-import com.gradeCalculator.server.Entities.ModuleEntity;
+import com.gradeCalculator.Entities.ModuleEntity;
 
 public interface ModuleRepository extends CrudRepository<ModuleEntity, Integer>{
     Iterable<ModuleEntity> findByName(String name);
