+++ src/main/java/com/gradeCalculator/repositories/UserRepository.java
@@ -1,6 +1,6 @@
-package com.gradeCalculator.server.repositories;
+package com.gradeCalculator.repositories;
 
-import com.gradeCalculator.server.Entities.UserEntity;
+import com.gradeCalculator.Entities.UserEntity;
 import org.springframework.data.repository.CrudRepository;
 
 public interface UserRepository extends CrudRepository<UserEntity, String> {
