+++ src/test/java/com/gradeCalculator/ServerApplicationTests.java
@@ -1,4 +1,4 @@
-package com.gradeCalculator.server;
+package com.gradeCalculator;
 
 import org.junit.jupiter.api.Test;
 import org.springframework.boot.test.context.SpringBootTest;
